codeclimate note fix
