+++ src/main/java/ru/mail/polis/lsm/artemdrozdov/LsmDAO.java
@@ -131,7 +131,7 @@ public class LsmDAO implements DAO {
             });
 
             compactExecutor.execute(() -> {
-                if (compactLock.tryLock() && tableStorage.isCompact()) {
+                if (compactLock.tryLock() && tableStorage.isCompact(config.tableLimit)) {
                     try {
                         compact();
                     } finally {
@@ -152,7 +152,7 @@ public class LsmDAO implements DAO {
         try {
             sizeBeforeCompact.set(tableStorage.tables.size());
             SSTable table = perfomCompact();
-            this.tableStorage = tableStorage.afterCompact(table);
+            this.tableStorage = tableStorage.afterCompact(table, sizeBeforeCompact.get());
         } catch (IOException e) {
             throw new UncheckedIOException("Can't compact", e);
         }
@@ -176,6 +176,8 @@ public class LsmDAO implements DAO {
 
     @Override
     public void close() throws IOException {
+        // после изменения CompletebleFeature.runAsync на ExecutorService,
+        // время выполнения тестов hugeRecord увеличилось
         flushExecutor.shutdown();
         compactExecutor.shutdown();
         try {
@@ -190,7 +192,7 @@ public class LsmDAO implements DAO {
             throw new IllegalStateException(e);
         }
         synchronized (this) {
-            if (tableStorage.isCompact()) {
+            if (tableStorage.isCompact(config.tableLimit)) {
                 compact();
             }
             flush(memoryStorage);
@@ -277,38 +279,5 @@ public class LsmDAO implements DAO {
                 return delegate.next();
             }
         };
-    }
-
-    private class TableStorage {
-        public final List<SSTable> tables;
-
-        TableStorage(final List<SSTable> newTables) {
-            this.tables = newTables;
-        }
-
-        TableStorage(final SSTable table) {
-            this(Collections.singletonList(table));
-        }
-
-        public TableStorage afterFlush(SSTable newTable) {
-            List<SSTable> newTables = new CopyOnWriteArrayList<>();
-            newTables.addAll(tables);
-            newTables.add(newTable);
-            return new TableStorage(newTables);
-        }
-
-        public TableStorage afterCompact(SSTable compactTable) {
-            List<SSTable> newTables = new CopyOnWriteArrayList<>();
-            // во время компакта, ещё флашились таблицы -> нужно их добавить
-            for (int i = sizeBeforeCompact.get(); i < tables.size(); ++i) {
-                newTables.add(tables.get(i));
-            }
-            newTables.add(compactTable);
-            return new TableStorage(newTables);
-        }
-
-        public boolean isCompact() {
-            return tables.size() >= config.tableLimit;
-        }
     }
 }
